Conflicts:
	Data/Lua/init.lua
	MasterUAB/Core/Core.vcxproj
	MasterUAB/Core/Core.vcxproj.filters
	MasterUAB/Core/Engine/UABEngine.cpp
	MasterUAB/Graphics/Manchas/ManchasSystemInstance.cpp

+++ MasterUAB/Graphics/Manchas/ManchasSystemInstance.h
@@ -24,7 +24,7 @@ private:
 		float Size;
 		float SizeSpeed;
 		float Opacity, MaxOpacity;
-		int Frame;
+		float Frame;
 		CColor Color;
 	};
